render_refactor split render_functions.py into many files, as it was waaay tooo laaarge documented and TYPEd the methods correctly
